healthcheck: Do not automatically invoke healthcheck
Running the docker healthcheck is an anti-pattern and 'frowned' upon, as
it takes away control from the user. A very simple relevant example is
the 6 minute startup delay. This is only true in certain conditions. If
a system is slower, the 6 minutes might need to be 10 minutes.

Because of this, it is recommended to add the needed scripts, but never
run them automatically from the Dockerfile.

Document how to do this instead and give control back to the user.

While here, also document how this would look like with docker compose.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>