Fix: Avoid running /init as pid 1
Running a shell script as pid 1 is to be avoided as
processes cannot be reapt.

Tini, a simple init system to ensure processes are properly handled can
do this for us, without a high cost.

The nice thing is also, this removes the diff between the various
entrypoints, reducing the complexity of handling so many similar
scripts.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>