socket: Move to dedicated directory
While using '/tmp' is fine for the socket, it makes things harder when
wanting to share the socket with docker volumes. If you mount it as a
file, it needs to already exists when creating the container, as
otherwise docker will create a directory with the name of the socket,
causing clamav to fail to create it. But volume-mounting /tmp is also
not desirable.

So moving the socket into its own directory yields the best of both
worlds.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>